+++ packages/terra-clinical-onset-picker/tests/wdio/onset-spec.js
@@ -1,173 +1,124 @@
 const viewports = Terra.viewports('tiny', 'medium', 'enormous');
-// Color contrast rule disabled due to https://github.com/cerner/terra-core/issues/1917
-const axeRules = { 'color-contrast': { enabled: false } };
 
 viewports.forEach((viewport) => {
-  describe(`Onset Picker [${viewport}]`, () => {
-    before(() => {
-      browser.setViewportSize(viewport);
-    });
+  describe(`Onset Picker [${viewport.name}]`, () => {
+    before(() => browser.setViewportSize(viewport));
 
-    describe('When unknown precision selected', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
-        browser.click('#test-precision-select');
-        browser.click('#terra-select-option-unknown');
-      });
+    describe('has [year granularity] select', () => {
+      before(() => browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default'));
 
-      Terra.should.validateElement({ axeRules });
-    });
-
-    // Year granularity
-    describe('Displays year selection only when Year granularity selected', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
+      it('displays year select', () => {
         browser.click('#test-granularity-select');
         browser.click('#terra-select-option-year');
       });
 
-      Terra.should.validateElement({ axeRules });
-    });
+      Terra.it.validatesElement();
 
-    describe('Can select a year between the birthdate and current year', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-year');
-
+      it('has year options', () => {
         browser.click('#test-year-select');
-        browser.click('#terra-select-option-2015');
-      });
-
-      Terra.should.validateElement({ axeRules });
-    });
-
-    describe('Cannot select a year before the birthdate', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-year');
-      });
-
-      it('does not show a year before birthdate', () => {
-        browser.click('#test-year-select');
-        browser.isExisting('#terra-select-option-2010').should.equal(false);
         // move the mouse to prevent mis-match with hover styling
-        browser.moveToObject('#terra-select-option-2011');
+        browser.moveToObject('#terra-select-option-2015');
       });
 
-      Terra.should.validateElement({ selector: '#root', axeRules });
-    });
+      Terra.it.validatesElement('year options', { selector: '#root' });
 
-    describe('Cannot select a year in the future', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-year');
+      it('does not show a year before birth date', () => {
+        browser.isExisting('#terra-select-option-2010').should.equal(false);
       });
 
       it('does not show a year in the future', () => {
-        browser.click('#test-year-select');
         browser.isExisting('#terra-select-option-2017').should.equal(false);
-        // move the mouse to prevent mis-match with hover styling
-        browser.moveToObject('#terra-select-option-2016');
       });
 
-      Terra.should.validateElement({ selector: '#root', axeRules });
-    });
-
-    // Monthly granularity
-    describe('Displays month and year selection only when Month-Year granularity selected', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-month');
-      });
-
-      Terra.should.validateElement({ axeRules });
-    });
-
-    describe('Can select a month between the birthdate and current date', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-month');
-
-        browser.click('#test-year-select');
+      it('can select a year between the birth date and current year', () => {
         browser.click('#terra-select-option-2015');
-
-        browser.click('#test-month-select');
-        browser.click('#terra-select-option-3');
       });
 
-      Terra.should.validateElement({ axeRules });
+      Terra.it.validatesElement('year option selected');
     });
 
-    describe('Cannot select a month before the birthdate', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
+    describe('has [month-year granularity] select', () => {
+      before(() => browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default'));
+
+      it('displays month and year selects', () => {
         browser.click('#test-granularity-select');
         browser.click('#terra-select-option-month');
+      });
 
+      Terra.it.validatesElement();
+
+      it('does not show a month before the birth date', () => {
         browser.click('#test-year-select');
         browser.click('#terra-select-option-2011');
-      });
-
-      it('does not show a month before the birthdate', () => {
         browser.click('#test-month-select');
         browser.isExisting('#terra-select-option-1').should.equal(false);
-        // move the mouse to prevent mis-match with hover styling
-        browser.moveToObject('#terra-select-option-8');
-      });
-
-      Terra.should.validateElement({ selector: '#root', axeRules });
-    });
-
-    describe('Cannot select a month in the future', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-month');
-
-        browser.click('#test-year-select');
-        browser.click('#terra-select-option-2016');
+        // close month select dropdown for tiny viewport
+        browser.click('#test-month-select');
       });
 
       it('does not show a month in the future', () => {
+        browser.click('#test-year-select');
+        browser.click('#terra-select-option-2016');
         browser.click('#test-month-select');
         browser.isExisting('#terra-select-option-9').should.equal(false);
-        // move the mouse to prevent mis-match with hover styling
-        browser.moveToObject('#terra-select-option-8');
+        // close month select dropdown for tiny viewport
+        browser.click('#test-month-select');
       });
 
-      Terra.should.validateElement({ selector: '#root', axeRules });
+      it('has month options', () => {
+        browser.click('#test-year-select');
+        browser.click('#terra-select-option-2015');
+        browser.click('#test-month-select');
+        // move the mouse to prevent mis-match with hover styling
+        browser.moveToObject('#terra-select-option-3');
+      });
+
+      Terra.it.validatesElement('month options', { selector: '#root' });
+
+      it('Can select a month between the birth date and current year', () => {
+        browser.click('#terra-select-option-3');
+      });
+
+      Terra.it.validatesElement('month option selected');
     });
 
-    // Date granularity
-    describe('Displays date input when Date granularity is selected', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
+    describe('has [date granularity]', () => {
+      before(() => browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default'));
+
+      it('displays date input', () => {
         browser.click('#test-granularity-select');
         browser.click('#terra-select-option-date');
       });
 
-      Terra.should.validateElement({ axeRules });
+      Terra.it.validatesElement();
     });
 
-    // Age granularity
-    describe('Displays age inputs when Age granularity is selected', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default');
+    describe('has [age granularity]', () => {
+      before(() => browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default'));
+
+      it('displays age inputs', () => {
         browser.click('#test-granularity-select');
         browser.click('#terra-select-option-age');
       });
 
-      Terra.should.validateElement({ axeRules });
+      Terra.it.validatesElement();
     });
 
-    describe('Cannot select number of years greater than age', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years');
+    describe('has [unknown precision]', () => {
+      before(() => browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/default'));
+
+      it('displays no onset options', () => {
+        browser.click('#test-precision-select');
+        browser.click('#terra-select-option-unknown');
+      });
+
+      Terra.it.validatesElement();
+    });
+
+    describe('age set to [five years]', () => {
+      before(() => browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years'));
+
+      it('cannot select number of years greater than age', () => {
         browser.click('#test-granularity-select');
         browser.click('#terra-select-option-age');
 
@@ -180,15 +131,9 @@ viewports.forEach((viewport) => {
         browser.pause(2900);
       });
 
-      Terra.should.validateElement({ selector: '#root', axeRules });
-    });
+      Terra.it.validatesElement('invalid year input', { selector: '#root' });
 
-    describe('Cannot select more than 24 months', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-age');
-
+      it('cannot select more than 24 months', () => {
         browser.click('#test-age-unit-select');
         browser.click('#terra-select-option-months');
 
@@ -198,33 +143,9 @@ viewports.forEach((viewport) => {
         browser.pause(2900);
       });
 
-      Terra.should.validateElement({ selector: '#root', axeRules });
-    });
+      Terra.it.validatesElement('invalid month input', { selector: '#root' });
 
-    describe('Cannot select number of months that exceede age', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/less-year');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-age');
-
-        browser.click('#test-age-unit-select');
-        browser.click('#terra-select-option-months');
-
-        browser.clearElement('#test-age-input');
-        browser.setValue('#test-age-input', 20);
-        browser.click('button[type="submit"]');
-        browser.pause(2900);
-      });
-
-      Terra.should.validateElement({ selector: '#root', axeRules });
-    });
-
-    describe('Cannot select more than 8 weeks', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-age');
-
+      it('cannot select more than 8 weeks', () => {
         browser.click('#test-age-unit-select');
         browser.click('#terra-select-option-weeks');
 
@@ -234,16 +155,45 @@ viewports.forEach((viewport) => {
         browser.pause(2900);
       });
 
-      Terra.should.validateElement({ selector: '#root', axeRules });
+      Terra.it.validatesElement('invalid week input', { selector: '#root' });
     });
 
-    describe('Cannot select number of weeks that exceede age', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/less-month');
+    describe('has age set to [less than one year] ', () => {
+      before(() => browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/less-year'));
+
+      it('cannot select years duration if age is less than a year old', () => {
         browser.click('#test-granularity-select');
         browser.click('#terra-select-option-age');
 
         browser.click('#test-age-unit-select');
+        browser.isExisting('#terra-select-option-years').should.equal(false);
+      });
+
+      it('cannot select number of months that exceed age', () => {
+        browser.click('#terra-select-option-months');
+
+        browser.clearElement('#test-age-input');
+        browser.setValue('#test-age-input', 20);
+        browser.click('button[type="submit"]');
+        browser.pause(2900);
+      });
+
+      Terra.it.validatesElement('invalid month input', { selector: '#root' });
+    });
+
+    describe('has age set to [less than one month] ', () => {
+      before(() => browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/less-month'));
+
+      it('cannot select months duration if age is less than a month old', () => {
+        browser.click('#test-granularity-select');
+        browser.click('#terra-select-option-age');
+
+        browser.click('#test-age-unit-select');
+        browser.isExisting('#terra-select-option-months').should.equal(false);
+        browser.isExisting('#terra-select-option-years').should.equal(false);
+      });
+
+      it('cannot select number of months that exceed age', () => {
         browser.click('#terra-select-option-weeks');
 
         browser.clearElement('#test-age-input');
@@ -252,123 +202,65 @@ viewports.forEach((viewport) => {
         browser.pause(2900);
       });
 
-      Terra.should.validateElement({ selector: '#root', axeRules });
+      Terra.it.validatesElement('invalid month input', { selector: '#root' });
     });
 
-    describe('Cannot select years duration if age is less than a year old', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/less-year');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-age');
-      });
+    describe('has [onsetOnChange callback]', () => {
+      before(() => browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years'));
 
-      it('does not show years if less than a year old', () => {
-        browser.click('#test-age-unit-select');
-        browser.isExisting('#terra-select-option-years').should.equal(false);
-      });
-
-      Terra.should.validateElement({ selector: '#root', axeRules });
-    });
-
-    describe('Cannot select months duration if age is less than a month old', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/less-month');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-age');
-      });
-
-      it('does not show months if less than a month old', () => {
-        browser.click('#test-age-unit-select');
-        browser.isExisting('#terra-select-option-months').should.equal(false);
-      });
-
-      Terra.should.validateElement({ selector: '#root', axeRules });
-    });
-
-    // onChange Handler
-    describe('When precision is changed an event is fired to the onsetOnChange callback', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years');
-        browser.click('#test-precision-select');
-        browser.click('#terra-select-option-unknown');
-        browser.click('button[type="submit"]');
-      });
-
-      Terra.should.validateElement({ axeRules });
-    });
-
-    describe('When granularity is changed an event is fired to the onsetOnChange callback', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-age');
-        browser.click('button[type="submit"]');
-      });
-
-      Terra.should.validateElement({ axeRules });
-    });
-
-    describe('When onsetDate is changed by the month select an event is fired to the onsetOnChange callback', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years');
+      it('that fires when granularity is changed', () => {
         browser.click('#test-granularity-select');
         browser.click('#terra-select-option-month');
+        browser.click('button[type="submit"]');
+      });
 
+      Terra.it.validatesElement('granularity changed');
+
+      it('that fires when month select is changed', () => {
         browser.click('#test-month-select');
         browser.click('#terra-select-option-9');
         browser.click('button[type="submit"]');
       });
 
-      Terra.should.validateElement({ axeRules });
-    });
+      Terra.it.validatesElement('month select changed');
 
-    describe('When onsetDate is changed by the year select an event is fired to the onsetOnChange callback', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years');
+      it('that fires when year select is changed', () => {
         browser.click('#test-granularity-select');
         browser.click('#terra-select-option-year');
-
         browser.click('#test-year-select');
         browser.click('#terra-select-option-2016');
         browser.click('button[type="submit"]');
       });
 
-      Terra.should.validateElement({ axeRules });
-    });
+      Terra.it.validatesElement('year select changed');
 
-    describe('When onsetDate is changed by the age input an event is fired to the onsetOnChange callback', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years');
-        browser.click('#test-granularity-select');
-        browser.click('#terra-select-option-age');
-
-        browser.click('#test-age-unit-select');
-        browser.click('#terra-select-option-years');
-
-        browser.clearElement('#test-age-input');
-        browser.setValue('#test-age-input', 4);
-        browser.click('button[type="submit"]');
-      });
-
-      Terra.should.validateElement({ axeRules });
-    });
-
-    describe('When onsetDate is changed by the ageUnit select an event is fired to the onsetOnChange callback', () => {
-      before(() => {
-        browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/five-years');
+      it('that fires when age input is changed', () => {
         browser.click('#test-granularity-select');
         browser.click('#terra-select-option-age');
 
         browser.clearElement('#test-age-input');
         browser.setValue('#test-age-input', 4);
 
-        browser.click('#test-age-unit-select');
-        browser.click('#terra-select-option-years');
         browser.click('button[type="submit"]');
-        browser.pause(2900);
       });
 
-      Terra.should.validateElement({ axeRules });
+      Terra.it.validatesElement('age input changed');
+
+      it('that fires when ageUnit select is changed', () => {
+        browser.click('#test-age-unit-select');
+        browser.click('#terra-select-option-weeks');
+        browser.click('button[type="submit"]');
+      });
+
+      Terra.it.validatesElement('ageUnit select changed');
+
+      it('that fires when precision is changed', () => {
+        browser.click('#test-precision-select');
+        browser.click('#terra-select-option-unknown');
+        browser.click('button[type="submit"]');
+      });
+
+      Terra.it.validatesElement('precision changed');
     });
 
     describe('Hides legend when isLegendHidden is true', () => {
@@ -376,7 +268,7 @@ viewports.forEach((viewport) => {
         browser.url('/#/raw/tests/terra-clinical-onset-picker/clinical-onset-picker/with-hidden-legend');
       });
 
-      Terra.should.validateElement({ axeRules });
+      Terra.it.validatesElement();
     });
   });
 });