+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added enumeration for valid column sort indicator values
 
+### Changed
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
+
 2.5.0 - (May 7, 2019)
 ----------
 ### Changed