+++ packages/terra-clinical-item-view/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added enumeration for valid layout, emphasis, and alignment values
 
+### Changed
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
+
 3.3.0 - (March 26, 2019)
 ----------
 ### Changed