+++ packages/terra-clinical-onset-picker/CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased
 * Added legend and isLegendHidden prop to Onsetpicker field group.
 * Added enumeration for valid age unit, granularity, precision, and precision set values
 
+### Changed
+* Update tests for terra-toolkit v5 and terra-dev-site v5 changes
+
 4.2.0 - (April 16, 2019)
 ----------
 ### Fixed